requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
autopep8==2.3.2
babel==2.18.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
datasets==2.18.0
decorator==5.2.1
dgl==1.1.2
dglgo==0.0.2
dill==0.3.8
docutils==0.21.2
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
greenlet==3.3.1
grpcio==1.78.0
hf-xet==1.2.0
huggingface_hub==0.36.2
idna==3.11
imagesize==1.4.1
isort==8.0.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.1
littleutils==0.2.4
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
networkx==3.4.2
numba==0.63.1
numpy==1.26.4
numpydoc==1.10.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
opencv-python==4.9.0.80
optuna==3.6.1
outdated==0.2.2
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.2.1
pandas-stubs==2.2.3.250527
pillow==10.2.0
platformdirs==4.8.0
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyarrow-hotfix==0.7
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyg-lib @ file:///home/li/%E4%B8%8B%E8%BD%BD/pyg_lib-0.4.0%2Bpt22cu121-cp310-cp310-linux_x86_64.whl#sha256=d6ee9fbf6588bae64f1a33834116e777b50528df401c23cdd4d6499b42814719
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.1
rdkit-pypi==2022.9.5
regex==2026.1.15
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
snowballstemmer==3.0.1
soundfile==0.13.1
soxr==1.0.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.46
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.15.2
tomli==2.4.0
torch==2.2.2+cu121
torch-geometric==2.7.0
torch_cluster @ file:///home/li/%E4%B8%8B%E8%BD%BD/torch_cluster-1.6.3%2Bpt22cu121-cp310-cp310-linux_x86_64.whl#sha256=055c78d49a34fd82219a29d24f55ee628487b6fe603ea4ea3c72d802fe8654da
torch_scatter @ file:///home/li/%E4%B8%8B%E8%BD%BD/torch_scatter-2.1.2%2Bpt22cu121-cp310-cp310-linux_x86_64.whl#sha256=caab7dd4a8783d49b2a983331e93ebf968c9c76ff4daaca158e2f6fb9c8028bf
torch_sparse @ file:///home/li/%E4%B8%8B%E8%BD%BD/torch_sparse-0.6.18%2Bpt22cu121-cp310-cp310-linux_x86_64.whl#sha256=67877b8d085a54b228e86ab52bcaeb0eceed11df19d686d3e25c46f4e7417837
torch_spline_conv @ file:///home/li/%E4%B8%8B%E8%BD%BD/torch_spline_conv-1.2.2%2Bpt22cu121-cp310-cp310-linux_x86_64.whl#sha256=11ea2cd645c58abc04270982ff97558b215766e95043cfb6452b6d32783efaa7
torchaudio==2.2.2+cu121
torchdata==0.7.1
torchvision==0.17.2+cu121
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typer==0.24.1
types-pytz==2025.2.0.20251108
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
